+++ src/libGLESv2/renderer/IndexRangeCache.h
@@ -18,9 +18,9 @@ namespace rx
 class IndexRangeCache
 {
   public:
-    void addRange(GLenum type, intptr_t offset, GLsizei count, unsigned int minIdx, unsigned int maxIdx, 
+    void addRange(GLenum type, unsigned int offset, GLsizei count, unsigned int minIdx, unsigned int maxIdx,
                   unsigned int streamOffset);
-    bool findRange(GLenum type, intptr_t offset, GLsizei count, unsigned int *outMinIndex,
+    bool findRange(GLenum type, unsigned int offset, GLsizei count, unsigned int *outMinIndex,
                    unsigned int *outMaxIndex, unsigned int *outStreamOffset) const;
 
     void invalidateRange(unsigned int offset, unsigned int size);
@@ -30,7 +30,7 @@ class IndexRangeCache
     struct IndexRange
     {
         GLenum type;
-        intptr_t offset;
+        unsigned int offset;
         GLsizei count;
 
         IndexRange();
